Remove the docs website tutorial include workaround
Previously, the asciidoc tutorial parts had to be included with liquid template syntax in the include statement. The tutorial index had to be split into two separate files because of that. This was due to the issue asciidoctor/jekyll-asciidoc#80.

Now the include issue was fixed, and the docs website was upgraded to have the fixed plugin. Therefore, the include workaround is no longer necessary.

In this change, the include workaround is removed, and the tutorial index is combined back into a single file.